Figures shifted (#519)
* figs

* shift figures

* figure 9 is on now

* linewidth for plottres
