+++ lineage/figures/figureCommon.py
@@ -118,7 +118,7 @@ def subplotLabel(axs):
             i += 1
 
 
-def overlayCartoon(figFile, cartoonFile, x, y, scalee=1, scale_x=1, scale_y=1):
+def overlayCartoon(figFile, cartoonFile, x, y, scalee=1, scale_x=1, scale_y=1, rotate=None):
     """ Add cartoon to a figure file. """
 
     # Overlay Figure cartoons
@@ -127,6 +127,8 @@ def overlayCartoon(figFile, cartoonFile, x, y, scalee=1, scale_x=1, scale_y=1):
 
     cartoon.moveto(x, y, scale=scalee)
     cartoon.scale_xy(scale_x, scale_y)
+    if rotate:
+        cartoon.rotate(rotate, x, y)
 
     template.append(cartoon)
     template.save(figFile)
