+++ makefile
@@ -2,7 +2,7 @@ SHELL := /bin/bash
 
 .PHONY: clean test testprofile testcover docs
 
-flist = 1 2 3 4 5 6 7 8 S01 S02 S03 S04 S05 S06 S07 S08 S09 S10 S12
+flist = 1 2 3 4 5 6 7 8 9 S01 S02 S03 S04 S05 S06 S07 S08 S09 S10 S12
 
 all: spell.txt $(patsubst %, output/figure%.svg, $(flist))
